requirements.txt.txt
pillow>=10.0.0
numpy>=1.24.0
pynput>=1.7.6
pytesseract>=0.3.10
requests>=2.31.0
pywin32>=306
comtypes>=1.2.0
psutil>=5.9.0
mss>=9.0.1
pyautogui>=0.9.54
packaging>=21.0